Make zimHttpServer32 compatible with AVM Fritz!Box routers (freetz)
AVM Fritz!Box products are big endian machines, while the integers
and shorts in the zim file format are stored in little endian byte
order.

Essentially, this patch adds robustness to read the zim file format,
regardless of the endianess of the target machine that executes perl
and this script.  For this, unpack("I") and unpack("s") statements
were replaced.

- Execution with microperl was not tested!

- Current freetz does not contain make/perl,
  but see https://github.com/dirk-dhu/freetz/tree/master/make/perl

- Building perl-5.26 in combination with cross-perl out of the
  freetz tree and integrating it manually to an external file
  is another possibility to get full perl support, see
  README.avm.perl

- Busybox (tested 1.24 and current 1.28) xz applet does not work
  (reports "corrupted data" to the terminal)

- xz from xz-utils works, but only in combination with swap
  enabled (xz fragments in zim files may have large dict size,
  expect the xz decompressor to use ~65mb to decompress; because
  of this 7270 and older boxes may not be suited to serve a zim)