rework xz detection
The busybox xz decompressor actually works if there is enough RAM available,
the error message ("corrupted data") it prints is misleading in low RAM env.

This means we do not need to care, if xz in system path is supplied by busybox
or xz-utils.  If there is no system supplied xz binary, take the local one.